Scrap GSL-based SR optimization and first round of polish
